+++ integrations/linear/webhook_processors/__init__.py
@@ -1,4 +1,4 @@
-from webhook_processors.issue_processor import IssueWebhookProcessor
-from webhook_processors.label_processor import LabelWebhookProcessor
+from webhook_processors.issue_webhook_processor import IssueWebhookProcessor
+from webhook_processors.label_webhook_processor import LabelWebhookProcessor
 
 __all__ = ["IssueWebhookProcessor", "LabelWebhookProcessor"]
